examples/label_studio_text_annotation/requirements.txt
pytorch_lightning
evaluate
transformers
datasets
scikit-learn
accelerate>=0.20.1